Changed publish methods to take advantage of the new IPM
