+++ rclcpp/include/rclcpp/intra_process_manager_impl.hpp
@@ -34,7 +34,7 @@
 
 #include "rclcpp/macros.hpp"
 #include "rclcpp/mapped_ring_buffer.hpp"
-#include "rclcpp/publisher.hpp"
+#include "rclcpp/publisher_base.hpp"
 #include "rclcpp/subscription.hpp"
 #include "rclcpp/visibility_control.hpp"
