+++ rclcpp/src/rclcpp/node_interfaces/node_topics.cpp
@@ -47,13 +47,9 @@ NodeTopics::create_publisher(
     // Get the intra process manager instance for this context.
     auto ipm = context->get_sub_context<rclcpp::intra_process_manager::IntraProcessManager>();
     // Register the publisher with the intra process manager.
-    uint64_t intra_process_publisher_id =
-      publisher_factory.add_publisher_to_intra_process_manager(ipm.get(), publisher);
-    // Create a function to be called when publisher to do the intra process publish.
-    auto shared_publish_callback = publisher_factory.create_shared_publish_callback(ipm);
+    uint64_t intra_process_publisher_id = ipm->add_publisher(publisher);
     publisher->setup_intra_process(
       intra_process_publisher_id,
-      shared_publish_callback,
       ipm,
       publisher_options);
   }
